feat(recherche): disable thematique + handle when no types
