+++ front/src/components/aroundMe/aroundMeFilter.component.tsx
@@ -47,7 +47,8 @@ const AroundMeFilter: React.FC<Props> = ({ visible, hideModal }) => {
     { title: string; filters: CartoFilter[] }[]
   >([]);
   const [cartoFilterStorage, setCartoFilterStorage] =
-    useState<CartoFilterStorage>({ etapes: [], thematiques: [], types: [] });
+    useState<CartoFilterStorage>({ etapes: [], types: [] });
+  // useState<CartoFilterStorage>({ etapes: [], thematiques: [], types: [] });
   const [showModalContent, setShowModalContent] = useState(false);
   const [trackerAction, setTrackerAction] = useState("");
 
@@ -128,10 +129,10 @@ const AroundMeFilter: React.FC<Props> = ({ visible, hideModal }) => {
           filters: fetchedFiltersFromDB.professionnels,
           title: Labels.aroundMe.filter.healthProfessional,
         },
-        {
-          filters: fetchedFiltersFromDB.etapes,
-          title: Labels.aroundMe.filter.steps,
-        },
+        // {
+        //   filters: fetchedFiltersFromDB.etapes,
+        //   title: Labels.aroundMe.filter.steps,
+        // },
       ]);
     }
   };
@@ -322,7 +323,7 @@ const AroundMeFilter: React.FC<Props> = ({ visible, hideModal }) => {
                 onPress={() => {
                   setCartoFilterStorage({
                     etapes: [],
-                    thematiques: [],
+                    // thematiques: [],
                     types: [],
                   });
                   hideModal(false);
@@ -361,7 +362,7 @@ const AroundMeFilter: React.FC<Props> = ({ visible, hideModal }) => {
                     action={() => {
                       setCartoFilterStorage({
                         etapes: [],
-                        thematiques: [],
+                        // thematiques: [],
                         types: [],
                       });
                       hideModal(false);
@@ -381,10 +382,10 @@ const AroundMeFilter: React.FC<Props> = ({ visible, hideModal }) => {
                       );
                       sendFiltersTracker(cartoFilterStorage.etapes);
                       sendFiltersTracker(cartoFilterStorage.types);
-                      sendFiltersTracker(cartoFilterStorage.thematiques);
+                      // sendFiltersTracker(cartoFilterStorage.thematiques);
                       setCartoFilterStorage({
                         etapes: [],
-                        thematiques: [],
+                        // thematiques: [],
                         types: [],
                       });
                       hideModal(true);
