+++ front/src/components/aroundMe/fetchPois.component.tsx
@@ -59,7 +59,7 @@ const FetchPois: React.FC<Props> = ({
       lat2: bottomRightPoint.latitude,
       long1: topLeftPoint.longitude,
       long2: bottomRightPoint.longitude,
-      thematiques: savedFilters?.thematiques ? savedFilters.thematiques : [],
+      // thematiques: savedFilters?.thematiques ? savedFilters.thematiques : [],
       types: savedFilters?.types ? savedFilters.types : [],
     };
     void getPoisByGpsCoords({
