+++ front/src/components/aroundMe/fetchPoisCoords.component.tsx
@@ -84,7 +84,7 @@ const FetchPoisCoords: React.FC<Props> = ({
       lat2: bottomRightPoint.latitude,
       long1: topLeftPoint.longitude,
       long2: bottomRightPoint.longitude,
-      thematiques: savedFilters?.thematiques ? savedFilters.thematiques : [],
+      // thematiques: savedFilters?.thematiques ? savedFilters.thematiques : [],
       types: savedFilters?.types ? savedFilters.types : [],
     };
     void getPoisByGpsCoords({
