+++ front/src/components/calendar/eventCard.component.tsx
@@ -61,7 +61,7 @@ const EventCard: FC<Props> = ({ event, isExpanded, onPressed }) => {
   const updateCartoFilterStorage = () => {
     const cartoFilterStorage: CartoFilterStorage = {
       etapes: _.map(event.etapes, "nom"),
-      thematiques: event.thematique?.nom ? [event.thematique.nom] : [],
+      // thematiques: event.thematique?.nom ? [event.thematique.nom] : [],
       types: [],
     };
     void StorageUtils.storeObjectValue(
