+++ front/src/components/search/tabSearchRoutes.component.tsx
@@ -1,4 +1,5 @@
-import type { FC, ReactElement } from "react";
+import type { PoiType } from "@socialgouv/nos1000jours-lib";
+import type { ReactElement } from "react";
 import * as React from "react";
 import { useState } from "react";
 import { StyleSheet, View } from "react-native";
@@ -7,9 +8,12 @@ import { ScrollView } from "react-native-gesture-handler";
 
 import ArticleCard from "../../components/article/articleCard.component";
 import { SecondaryTextItalic } from "../../components/baseComponents";
+import { Labels, StorageKeysConstants } from "../../constants";
 import ArticleDetail from "../../screens/articles/articleDetail.component";
 import { Margins, Paddings } from "../../styles";
+import type { CartoFilterStorage } from "../../type";
 import type { Article, Step } from "../../types";
+import { storeObjectValue } from "../../utils/storage.util";
 import TabAroundMeInstruction from "./tabAroundMeInstruction.component";
 
 export const articlesRoute = (
@@ -65,14 +69,50 @@ export const articlesRoute = (
 export const poisRoute = (
   updatedText: string,
   articles: Article[]
-): ReactElement =>
-  articles.length > 0 ? (
-    <TabAroundMeInstruction articles={articles} />
+): ReactElement => {
+  let searchCanBeLaunched = true;
+  if (articles.length > 0) {
+    const types = extractedPoiTypesFromArticles(articles);
+    if (types.length === 0) {
+      searchCanBeLaunched = false;
+      updatedText = Labels.search.cantLaunchAroundMeSearch;
+    }
+  }
+
+  return articles.length > 0 && searchCanBeLaunched ? (
+    <TabAroundMeInstruction />
   ) : (
     <View style={styles.center}>
       <SecondaryTextItalic>{updatedText}</SecondaryTextItalic>
     </View>
   );
+};
+
+const extractedPoiTypesFromArticles = (articles: Article[]) => {
+  const finalCartographieTypes: PoiType[] = [];
+  articles.forEach((article) => {
+    if (!article.cartographie_pois_types) return;
+    const filteredTypes = article.cartographie_pois_types.filter(
+      (type) =>
+        !finalCartographieTypes.some((finalType) => finalType.nom === type.nom)
+    );
+    finalCartographieTypes.push(...filteredTypes);
+  });
+
+  if (finalCartographieTypes.length > 0) {
+    const cartoFilterStorage: CartoFilterStorage = {
+      etapes: [],
+      // thematiques: [],
+      types: finalCartographieTypes.map((type) => type.nom),
+    };
+    void storeObjectValue(
+      StorageKeysConstants.cartoFilterKey,
+      cartoFilterStorage
+    );
+  }
+
+  return finalCartographieTypes;
+};
 
 const styles = StyleSheet.create({
   center: {
